feat(HyperbolicEdgeBundles): Better hover+focus.
Store more information in the field hover provider's state
so that components can specify 3 levels of interest: no interest,
highlight-only, or highlight+focus. The latter accepts a single
named field to become the "subject" of a component's visualization.
Many fields may be simultaneously highlighted, with or without a
focus.

Also, store the "disposition" of the hover state: is it part of
a single, long-running user interaction (in which case all but the
last state is marked 'preliminary') or is it a simple hover
motion (in which case the disposition is 'final')?